refactor(orion): remove impor cycles
